ssh: allow the client public key auth method to use custom algorithms during the ssh handshake.
To accomplish this, we create  a new "AlgorithmSignerWithAlgoName" interface that returns the custom algorithm being used for signing. This allows publicKeyCallback to handshake using the correct key algorihm in situations where key type != key algorithm. For example, ssh-rsa keys may need to handshake using the "rsa-sha2-256" algorithm if using a recent version of openssh that has deprecated the ssh-rsa sha1 algorithm.

Fixes #39885